Change the layout of the download pages to make them more user-friendly.
Estimated hours taken: 6

Change the layout of the download pages to make them more user-friendly.

The main download page now contains a short description of each available
version of the compiler and a link to a more detailed download page.
There is no longer a distinction between the stable and unstable releases of
the day on the website, instead we link to the latest ``unstable'' version
only.  The main download page contains a link to the current version and the
release of the day and a link to the beta version only if there is one.

The snapshot page now looks similar to the current release page, except that it
contains information about the release of the day.  Previously the snapshot
page contained information about all the different versions of the Mercury
compiler.

The latest release of the day version is determined by examining the files
currently on the web server.  Once the website has been installed, old versions
of the rotd are deleted from the website (this used to be done by the
test_mercury script which meant that it had to generate the html page
with links to the ROTD files).

w3/Makefile:
	Add a WWW_DOMAIN variable which can be used to test the website from
	another web server.

	Make the images directory executable by all.

	Add a new rule that removes old versions of the ROTD.
w3/Makefile.common:
	Give default values for WWW_DOMAIN and allow
	the php command to be overridden if necessary.

	Define BETA_FTPDIR and BETA_WEBDIR variables
	which point to where the rotd files are kept.

w3/RELEASE_INFO:
	Make NEXT_BETA_VERSION blank.  If NEXT_BETA_VERSION is set then there
	will be link to it from the main download page.

w3/bin/install_webpages
	Remove the old versions of the ROTD after the website has been
	installed.

w3/download/Makefile:
	Fix an incorrect version number.

w3/download/include/release.inc:
	Use the new php variables which contain the www and ftp domains.
	Remove a redundant comment.

w3/download/include/rotd.inc:
	Change the snapshot download page so that it looks more like the
	current release download page and links to the source
	distribution files directly.

w3/include/download.inc:
	Change the main download page to contain a description of each
	available version.
	Only show a beta version if there is a beta version available.

w3/include/globals.inc.in:
	Add variables for the ROTD version and the ftp and www domains.