Fix the search mailing lists function on the Mercury website.
Earlier I submitted a patch to fix the search function on the website since the
URL had changed.  I neglected to do the same for searching the mailing list
archives.  This patch corrects this and places a note in both locations to warn
developers against the same mistake.

Also update the domain name parts of the mailing list addresses.

include/search.inc:
    Include warning pointing developers to mailing-lists/include/mailing-lists.inc

mailing-lists/include/mailing-lists.inc:
    Update URLs in the search forms.

    Update Mailing list URLs.

    Include warning pointing developers to include/search.inc