Update bugtracker and mailing list URLs.
We've purchased an SSL certificate for www.mercurylang.org (it also works
for mercurylang.org).  Previously we had self-signed certificates for the
bugtracker and mailing list virtual hosts.  To make the one (more secure)
certificate work for the bugtracker and mailing list I've moved them to new
URLs.

    https://www.mercurylang.org/bugs
    https://www.mercurylang.org/lists

The mail archives are now at the following URL, which does not use HTTPS so
that proxy servers can still cache this content.

    http://www.mercurylang.org/list-archives

I've added redirects for all the old URLs so that users' browsers will
automatically navigate to the new URLs.  The Google bot and other search
engine crawlers will also follow these to understand that our content has
moved, this is important for the mailing list archives.

This patch updates these URLs on the website.

include/menubar.inc:
    Update the bugtracker's URL

include/contact.inc:
    Update mailing lists' and bugtracker's URLs.