Save the small versions of the Mercury logo
This logo was designed by Yes Logic http://www.yeslogic.com/ who designed
our new website.  The small version of the logo has the same design as the
large one (part of the website) but is square and useful as an avatar or
icon.  It is used on our twitter account and github project.

This logo is available in multiple resolutions and saved as part of the
website.

images/mercurylogo_110x110.png:
images/mercurylogo_110x110.xcf:
images/mercurylogo_110x80.png:
images/mercurylogo_110x80.xcf:
images/mercurylogo_376_376.png:
    Save the logo.