Update contributions page and the list of developers.
Copyright assignment is unnecessary and impractical for Mercury.  We have
only inconsistently applied this policy in the past.  Additionally
relicensing is intractable at this point and therefore copyright assignment
is not useful, and only serves to confuse potential developers.

This change removes the discussion of copyright assignment and relicensing
from Mercury's website.  Now this page states that we simply maintain a list
of all the developers and their employers (where applicable).

Also remove references to public domain as it's legally ambiguous (highly
dependent on jurisdiction).

development/include/contributions.inc:
    As above.